fix(security): change setup for role base controller

+++ src/main/java/org/example/dto/proposal/ProposalViewDto.java
@@ -1,10 +1,10 @@
 package org.example.dto.proposal;
 
 public record ProposalViewDto(
-    Long id,
-    Long expertId,
-    String firstName,
-    String lastName,
-    Integer expertScore,
-    Double proposedPrice
-){}
+        Long id,
+        Long expertId,
+        String firstName,
+        String lastName,
+        Double expertScore,
+        Double proposedPrice
+) {}
